# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-PERSON ",
        "description": "Documentação da API-PERSON com Swagger Service.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/pessoas": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Lista todos as pessoas",
                "description": "Retorna uma lista de pessoas armazenados no banco de dados.",
                "operationId": "ac423fa98d261d82297b099cf6195b5f",
                "responses": {
                    "200": {
                        "description": "Lista de pessoas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pessoas"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum pessoa encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não há pessoas cadastrados!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar os pessoas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os pessoas."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Cria uma nova pessoa",
                "description": "Registra uma nova pessoa no banco de dados.",
                "operationId": "53dddbadc4166ae9a5bf1b92ec4641ab",
                "requestBody": {
                    "description": "Dados necessários para criar uma nova pessoa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_nome",
                                    "pes_cpf",
                                    "pes_data_nascimento",
                                    "pes_sexo",
                                    "pes_mae",
                                    "pes_pai"
                                ],
                                "properties": {
                                    "pes_nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "pes_cpf": {
                                        "type": "string",
                                        "example": "111.222.333-44"
                                    },
                                    "pes_data_nascimento": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "1978-08-23"
                                    },
                                    "pes_sexo": {
                                        "type": "string",
                                        "example": "Masculino"
                                    },
                                    "pes_mae": {
                                        "type": "string",
                                        "example": "Maria Aparecida da Silva"
                                    },
                                    "pes_pai": {
                                        "type": "string",
                                        "example": "Cícero Joaquim da Silva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pessoa criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pessoas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o pessoa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o pessoa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pessoas": {
                "title": "Pessoa",
                "description": "Modelo de Pessoa",
                "required": [
                    "pes_nome",
                    "pes_cpf",
                    "pes_data_nascimento",
                    "pes_sexo",
                    "pes_mae",
                    "pes_pai"
                ],
                "properties": {
                    "pes_id": {
                        "description": "ID do Pessoa",
                        "type": "integer",
                        "example": 1
                    },
                    "pes_nome": {
                        "description": "Nome da pessoa",
                        "type": "string",
                        "example": "João da Silva"
                    },
                    "pes_cpf": {
                        "description": "CPF da pessoa",
                        "type": "string",
                        "example": "111.222.333-44"
                    },
                    "pes_data_nascimento": {
                        "description": "Data de nascimento",
                        "type": "string",
                        "format": "date-time",
                        "example": "1978-08-23"
                    },
                    "pes_sexo": {
                        "description": "Sexo da pessoa",
                        "type": "string",
                        "example": "Masculino"
                    },
                    "pes_mae": {
                        "description": "Nome da mãe",
                        "type": "string",
                        "example": "Maria Aparecida da Silva"
                    },
                    "pes_pai": {
                        "description": "Nome da pai",
                        "type": "string",
                        "example": "Cícero Joaquim da Silva"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}